requirements.txt
openai
streamlit
langchain
python-dotenv